+++ sulong/projects/com.oracle.truffle.llvm.nodes/src/com/oracle/truffle/llvm/nodes/intrinsics/sulong/LLVMPrintStackTrace.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2018, Oracle and/or its affiliates.
+ * Copyright (c) 2017, 2019, Oracle and/or its affiliates.
  *
  * All rights reserved.
  *
@@ -34,6 +34,7 @@ import java.util.List;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.TruffleException;
+import com.oracle.truffle.api.TruffleStackTrace;
 import com.oracle.truffle.api.TruffleStackTraceElement;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.nodes.Node;
@@ -68,8 +69,7 @@ public abstract class LLVMPrintStackTrace extends LLVMIntrinsic {
 
     private static SulongStackTrace getStackTrace(LLVMNode node, String message, boolean filterCurrentLocation) {
         Throwable t = new CThrowable(node, message);
-        TruffleStackTraceElement.fillIn(t);
-        List<TruffleStackTraceElement> ctrace = TruffleStackTraceElement.getStackTrace(t);
+        List<TruffleStackTraceElement> ctrace = TruffleStackTrace.getStacktrace(t);
 
         SulongStackTrace trace = new SulongStackTrace(message);
         for (int i = 0; i < ctrace.size(); i++) {